refactor(runtime_bundle): 提取 bundle 解包逻辑到 extraction 子模块
